chore: fix broken links and move link checker workflow to schedule (#469)
The link checker has been failing in a number of PRs lately. I found one link was actually broken because it was moved within denoland/deno, while other links are a bit flaky: broken in CI but available when visited in a browser. Therefore, as a suggestion I have moved the link checker to a separate workflow that runs on a schedule daily. This way it does not get in the way of PRs.